requirements.txt
beautifulsoup4==4.9.1
geopy==2.0.0
lxml==4.5.2
pymongo==3.11.0
python-dotenv==0.15.0
requests==2.22.0
urllib3==1.25.8
Unidecode==1.1.1
dnspython==1.16.0
APScheduler==3.0.0